Fast-play probability

+++ cpp/src/uct/UCTNode.hpp
@@ -305,7 +305,7 @@ public:
                 if(v < child_Q(action)) {
                     inverse_N[action] = 0.0f;
                 } else {
-                    inverse_N[action] = std::max(0.0f, m_uWeight * child_P_model(action) * sqrt(total_N) / (v - child_Q(action)) - 1);
+                    inverse_N[action] = std::max(0.0f, m_uWeight * child_P_model(action) * sqrtf(total_N) / (v - child_Q(action)) - 1);
                 }
                 sum += inverse_N[action];
             }
